Conexion y obtencion de datos --OK

+++ Dominio/Articulo.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Data.SqlTypes;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -12,13 +13,13 @@ namespace Dominio
 
         public string Codigo { get; set; }
 
-        public  string Nombre { get; set; }
+        public string Nombre { get; set; }
 
         public string Descripcion { get; set; }
 
         public string ImagenUrl { get; set; }
 
-        public float Precio { get; set; }
+        public decimal Precio { get; set; }
 
         public Marca Marca { get; set; }
